requirements.txt
pytest
allure-pytest
allure-python-commons
assertpy
python-dotenv
selenium
pytest-selenium
pytest-html-reporter
pytest-html